fix: remove mobile build targets from playground app
