+++ build-logic/convention/src/main/kotlin/org/publicvalue/convention/config/Kotlin.kt
@@ -2,19 +2,15 @@ package org.publicvalue.convention.config
 
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.configure
+import org.jetbrains.kotlin.gradle.ExperimentalKotlinGradlePluginApi
 import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
 
+@OptIn(ExperimentalKotlinGradlePluginApi::class)
 fun Project.configureKotlin() {
     configureJava()
     extensions.configure<KotlinMultiplatformExtension> {
-        targets.all {
-            compilations.all {
-                compileTaskProvider.configure {
-                    compilerOptions {
-                        freeCompilerArgs.add("-Xexpect-actual-classes")
-                    }
-                }
-            }
+        compilerOptions {
+//            freeCompilerArgs.add("-Xexpect-actual-classes")
         }
     }
 }
