+++ oidc-appsupport/src/commonMain/kotlin/org/publicvalue/multiplatform/oidc/appsupport/PlatformCodeAuthFlow.kt
@@ -1,5 +1,6 @@
 package org.publicvalue.multiplatform.oidc.appsupport
 
+import org.publicvalue.multiplatform.oidc.OpenIdConnectClient
 import org.publicvalue.multiplatform.oidc.flows.AuthCodeResponse
 import org.publicvalue.multiplatform.oidc.flows.CodeAuthFlow
 import org.publicvalue.multiplatform.oidc.flows.EndSessionFlow
@@ -11,4 +12,5 @@ expect class PlatformCodeAuthFlow: CodeAuthFlow, EndSessionFlow {
     // in kotlin 2.0, we need to implement methods in expect classes
     override suspend fun getAuthorizationCode(request: AuthCodeRequest): AuthCodeResponse
     override suspend fun endSession(request: EndSessionRequest): EndSessionResponse
+    override val client: OpenIdConnectClient
 }
