+++ oidc-tokenstore/build.gradle.kts
@@ -4,12 +4,14 @@ import org.publicvalue.convention.config.configureWasmTarget
 plugins {
     id("org.publicvalue.convention.android.library")
     id("org.publicvalue.convention.kotlin.multiplatform")
+    id("org.publicvalue.convention.kotlin.multiplatform.mobile")
     id("org.publicvalue.convention.centralPublish")
 }
 
 description = "Kotlin Multiplatform OIDC tokenstore library"
 
 kotlin {
+    jvm()
     configureIosTargets()
     configureWasmTarget()
     sourceSets {
