+++ playground-app/common/ui/resources/strings/build.gradle.kts
@@ -1,9 +1,9 @@
 plugins {
-    id("org.publicvalue.convention.android.library")
     id("org.publicvalue.convention.kotlin.multiplatform")
 }
 
 kotlin {
+    jvm()
     sourceSets {
         val commonMain by getting {
             dependencies {
