+++ playground-app/domain/src/commonMain/kotlin/org/publicvalue/multiplatform/oauth/domain/types/AuthorizeResult.kt
@@ -1,0 +1,14 @@
+package org.publicvalue.multiplatform.oauth.domain.types
+
+import org.publicvalue.multiplatform.oidc.types.AuthCodeRequest
+
+sealed class AuthorizeResult {
+    data class Request(
+        val authCodeRequestUrl: String,
+        val authCodeRequest: AuthCodeRequest
+    ): AuthorizeResult()
+    data class Response(
+        val authCode: String,
+        val authCodeResponseQueryString: String
+    ): AuthorizeResult()
+}
