+++ playground-app/shared/build.gradle.kts
@@ -1,17 +1,13 @@
 import org.publicvalue.convention.addKspDependencyForAllTargets
 
 plugins {
-    id("org.publicvalue.convention.android.library")
     id("org.publicvalue.convention.kotlin.multiplatform")
     id("org.publicvalue.convention.compose.multiplatform")
     alias(libs.plugins.ksp)
 }
 
-android {
-    sourceSets["main"].resources.srcDirs("src/androidMain/res", "src/commonMain/resources") // include resources in android
-}
-
 kotlin {
+    jvm()
     sourceSets {
         val commonMain by getting {
             dependencies {
@@ -31,13 +27,6 @@ kotlin {
                 api(projects.ui.idplist)
                 api(projects.ui.clientlist)
                 api(projects.ui.clientdetail)
-            }
-        }
-        val androidMain by getting {
-            dependencies {
-                api(libs.androidx.activity.compose)
-                api(libs.androidx.core.ktx)
-                api(libs.androidx.appcompat)
             }
         }
