+++ playground-app/shared/src/commonMain/kotlin/org/publicvalue/multiplatform/oauth/inject/CommonApplicationComponent.kt
@@ -1,17 +1,15 @@
 package org.publicvalue.multiplatform.oauth.inject
 
-import org.publicvalue.multiplatform.oauth.util.DispatcherProvider
-import org.publicvalue.multiplatform.oauth.util.DispatcherProviderImpl
-import org.publicvalue.multiplatform.oauth.logging.Logger
-import org.publicvalue.multiplatform.oauth.logging.StdoutLogger
 import me.tatarka.inject.annotations.Provides
 import org.publicvalue.multiplatform.oauth.data.inject.SqlDelightDatabaseComponent
-import org.publicvalue.multiplatform.oauth.domain.inject.WebserverComponent
+import org.publicvalue.multiplatform.oauth.logging.Logger
+import org.publicvalue.multiplatform.oauth.logging.StdoutLogger
+import org.publicvalue.multiplatform.oauth.util.DispatcherProvider
+import org.publicvalue.multiplatform.oauth.util.DispatcherProviderImpl
 
 interface CommonApplicationComponent:
     SqlDelightDatabaseComponent,
-    NoOpAppInitializerComponent,
-    WebserverComponent
+    NoOpAppInitializerComponent
 {
     val initializers: Set<AppInitializer>
