+++ playground-app/shared/src/jvmMain/kotlin/org/publicvalue/multiplatform/oauth/inject/DesktopApplicationComponent.kt
@@ -1,11 +1,12 @@
 package org.publicvalue.multiplatform.oauth.inject
 
 import me.tatarka.inject.annotations.Component
+import org.publicvalue.multiplatform.oauth.domain.inject.WebserverComponent
 
 @Component
 @ApplicationScope
 abstract class DesktopApplicationComponent(
-) : CommonDesktopApplicationComponent {
+) : CommonDesktopApplicationComponent, WebserverComponent {
 
     companion object
 }
