+++ sample-app/settings/build.gradle.kts
@@ -4,9 +4,11 @@ import org.publicvalue.convention.config.configureWasmTarget
 plugins {
     id("org.publicvalue.convention.android.library")
     id("org.publicvalue.convention.kotlin.multiplatform")
+    id("org.publicvalue.convention.kotlin.multiplatform.mobile")
 }
 
 kotlin {
+    jvm()
     sourceSets {
         configureIosTargets()
         configureWasmTarget()
